Add Bits Service's buildpack cache
[#115151789]

+++ app/jobs/runtime/buildpack_cache_cleanup.rb
@@ -6,6 +6,8 @@ module VCAP::CloudController
           logger = Steno.logger('cc.background')
           logger.info('Attempting cleanup of buildpack_cache blobstore')
 
+          return bits_client.delete_all_buildpack_caches if use_bits_service?
+
           blobstore = CloudController::DependencyLocator.instance.buildpack_cache_blobstore
           blobstore.delete_all
         end
@@ -16,6 +18,16 @@ module VCAP::CloudController
 
         def max_attempts
           3
+        end
+
+        private
+
+        def use_bits_service?
+          !!bits_client
+        end
+
+        def bits_client
+          CloudController::DependencyLocator.instance.bits_client
         end
       end
     end
